% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_delta_method.R
\name{delta_method_ATE}
\alias{delta_method_ATE}
\title{Delta Method for Average Treatment Effect (ATE)}
\usage{
delta_method_ATE(
  output,
  control_level = min(output$treatment),
  treatment_levels = setdiff(output$treatment, control_level)
)
}
\arguments{
\item{output}{The output object from a survival analysis using `survinf`.}

\item{control_level}{The level of the control group in the treatment vector.}

\item{treatment_levels}{The levels of the treatment groups to perform inference for. Default is all unique treatment levels except the control level.}
}
\value{
A modified output object with transformed estimates, standard errors, confidence intervals, and EIF for the ATE.
}
\description{
This function provides inference for the Average Treatment Effect (ATE) using the delta method.
}
