% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_delta_method.R
\name{delta_method_contrast}
\alias{delta_method_contrast}
\title{Delta Method Custom Function}
\usage{
delta_method_contrast(
  output,
  transform = function(est_treatment, est_control) {
     est_treatment - est_control

    },
  transform_grad = function(est_treatment, est_control) {
     c(1, -1)
 },
  control_level = min(output$treatment),
  treatment_levels = setdiff(output$treatment, control_level),
  name = "custom"
)
}
\arguments{
\item{output}{The output object from a survival analysis using `survinf`.}

\item{transform}{A function that transforms the treatment and control estimates. Default is the difference between treatment and control estimates.}

\item{transform_grad}{A function that computes the gradient of the transformation function. Default computes the gradient for difference in estimates.}

\item{control_level}{The level of the control group in the treatment vector. Default is the minimum value of the treatment vector.}

\item{treatment_levels}{The levels of the treatment groups to perform inference for. Default is all unique treatment levels except the control level.}

\item{name}{An optional name for the transformation being applied. Default is "custom".}
}
\value{
A modified output object with transformed estimates, standard errors, confidence intervals, and EIF (Efficient Influence Function) for the specified transformation.
}
\description{
This function provides inference for functions of the survival estimates outputted by `survinf` using the delta method.
}
