% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_delta_method.R
\name{delta_method_LRR}
\alias{delta_method_LRR}
\title{Delta Method for Log Relative Risk (LRR)}
\usage{
delta_method_LRR(
  output,
  control_level = min(output$treatment),
  treatment_levels = setdiff(output$treatment, control_level),
  exp_transform = TRUE
)
}
\arguments{
\item{output}{The output object from a survival analysis using `survinf`.}

\item{control_level}{The level of the control group in the treatment vector.}

\item{treatment_levels}{The levels of the treatment groups to perform inference for. Default is all unique treatment levels except the control level.}

\item{exp_transform}{Logical, indicating whether to exponentiate the transformed estimates and CI. Default is TRUE.}
}
\value{
A modified output object with transformed estimates, standard errors, confidence intervals, and EIF for the LRR.
}
\description{
This function provides inference for the Log Relative Risk (LRR) using the delta method.
}
