% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_debias_ftype_distr.R
\name{.target_event_type_distr}
\alias{.target_event_type_distr}
\title{internal use}
\usage{
.target_event_type_distr(
  target_failure_time,
  target_treatment,
  target_event_type,
  data_long,
  treatment_levels,
  event_type_levels,
  treatment.hat,
  censor.hazard.hats,
  total.hazard.hats,
  event_type.distr.hats,
  weights
)
}
\description{
internal use
}
