% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discrete_debias_failure_hazard.R
\name{.target_failure_hazard}
\alias{.target_failure_hazard}
\title{internal use}
\usage{
.target_failure_hazard(
  max_eps = 0.05,
  target_failure_time,
  target_treatment,
  target_event_type,
  data_long,
  treatment_levels,
  event_type_levels,
  treatment.hat,
  censor.hazard.hats,
  total.hazard.hats,
  event_type.distr.hats,
  weights,
  tol,
  verbose = F
)
}
\description{
internal use
}
