% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers_survival_data.R
\name{long_hazard_to_survival_mats}
\alias{long_hazard_to_survival_mats}
\title{internal use}
\usage{
long_hazard_to_survival_mats(
  hazard.hats,
  time_grid,
  left_cont = FALSE,
  as_vector = TRUE
)
}
\description{
internal use
}
