% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphthresh.R
\name{graphthresh}
\alias{graphthresh}
\title{Function to plot estimates and confidence intervals across different thresholds}
\usage{
graphthresh(res, monotone = F)
}
\arguments{
\item{res}{Dataframe obtained from running thresholdSurv() function}

\item{monotone}{TRUE to show the results with the monotonicity assumption}
}
\value{
A ggplot2 graph
}
\description{
Function to plot estimates and confidence intervals across different thresholds
}
