% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphthresh.R
\name{graphthresh}
\alias{graphthresh}
\title{Plot threshold-response estimates}
\usage{
graphthresh(res, monotone = F)
}
\arguments{
\item{res}{A \code{data.frame} obtained from the \code{thresholdSurv()} function, containing the estimates and related statistics.}

\item{monotone}{A \code{logical} value. Set to \code{TRUE} to display results under the monotonicity assumption.}
}
\value{
A \code{ggplot2} object displaying the estimates and confidence intervals for \eqn{E[E[Y|A \ge v, W]]} across the specified thresholds.
}
\description{
This function creates a plot of estimates and confidence intervals for \eqn{E[E[Y|A \ge v, W]]} across various thresholds.
}
