% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdBinary.R
\name{thresholdBinary}
\alias{thresholdBinary}
\title{Non-parametric estimation of the covariate-adjusted threshold-response function}
\usage{
thresholdBinary(
  data,
  covariates,
  outcome,
  marker,
  Delta = NULL,
  weights = NULL,
  threshold_list = NULL,
  sl_library = c("SL.mean", "SL.glm"),
  method = "method.CC_nloglik",
  cvControl = list(V = 5)
)
}
\arguments{
\item{data}{A \code{data.frame} containing the dataset to be analyzed.}

\item{covariates}{A \code{character} vector specifying the names of columns to be used as adjustment covariates.}

\item{outcome}{A \code{character} string specifying the name of the binary outcome column.}

\item{marker}{A \code{character} string specifying the name of the column representing the continuous marker.}

\item{Delta}{A \code{character} string specifying the name of the column indicating whether the outcome is observed.
If \code{NULL}, all outcome data is assumed to be available.}

\item{weights}{A \code{character} string specifying the name of the column containing the observation weights.
If \code{NULL}, all data are weighted equally.}

\item{threshold_list}{A \code{numeric} vector of threshold values for estimation.
If \code{NULL}, estimation is performed using the minimum value of the marker.}

\item{sl_library}{A \code{character} vector specifying the Super Learner librarIes to be used for estimation.
The default is \code{c("SL.mean", "SL.glm")}.}

\item{method}{A \code{character} string specifying the method used by Super Learner. The default is \code{"method.CC_nloglik"}.}

\item{cvControl}{A list specifying the cross-validation control parameters for Super Learner. The default is \code{list(V = 5)}.}
}
\value{
A \code{data.frame} containing the following columns for each threshold in \code{threshold_list}:
\itemize{
  \item \code{threshold}: The threshold value.
  \item \code{estimate}: The estimated value at the threshold.
  \item \code{se}: The standard error of the estimate.
  \item \code{ci_lo}: The lower bound of the confidence interval.
  \item \code{ci_hi}: The upper bound of the confidence interval.
  \item \code{estimate_monotone}: The monotone-adjusted estimate.
  \item \code{ci_lo_monotone}: The lower bound of the monotone-adjusted confidence interval.
  \item \code{ci_hi_monotone}: The upper bound of the monotone-adjusted confidence interval.
  \item \code{n_in_bin}: The number of observations above the threshold.
  \item \code{n_events_in_bin}: The number of events above the threshold.
}
}
\description{
This function estimates the covariate-adjusted probability of a binary outcome above
a range of thresholds for a continuous marker.
}
