% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdSurv.R
\name{thresholdSurv}
\alias{thresholdSurv}
\title{Non-parametric estimation of the covariate-adjusted threshold-response function
adjusting for right-censoring}
\usage{
thresholdSurv(
  data,
  covariates,
  failure_time,
  event_type,
  marker,
  tf,
  weights = NULL,
  threshold_list = NULL,
  nbins_time = 20,
  verbose = FALSE,
  learner.treatment = NULL,
  learner.event_type = NULL,
  learner.failure_time = NULL,
  learner.censoring_time = NULL,
  placebo_risk = NULL,
  placebo_se = NULL,
  modify_left_point = F,
  left_estimate = NA,
  left_se = NA
)
}
\arguments{
\item{data}{A \code{data.frame} containing the dataset to be analyzed.}

\item{covariates}{A \code{character} vector specifying the names of columns to be used as adjustment covariates.}

\item{failure_time}{A \code{character} string indicating the name of the column representing the failure time variable.}

\item{event_type}{A \code{character} string specifying the name of the column representing the event type variable.}

\item{marker}{A \code{character} string for the name of the column representing the marker variable.}

\item{tf}{A \code{numeric} value indicating the reference time point for the analysis.}

\item{weights}{A \code{character} string specifying the name of the column containing the weights for estimation.}

\item{threshold_list}{A \code{numeric} vector of threshold values for estimation. If \code{NULL}, estimation is performed on the entire dataset.}

\item{nbins_time}{A \code{numeric} value specifying the number of time bins to be used.}

\item{verbose}{A logical value indicating whether to display progress and diagnostic messages during the computation. Default is FALSE.}

\item{learner.treatment}{A \code{binomial} \code{\link[sl3]{Lrnr_base}} learner object used for fitting the propensity score model for the treatment mechanism.}

\item{learner.event_type}{A \code{binomial} \code{\link[sl3][Lrnr_base]} learner object used for fitting the conditional probability distribution for failure time of the failure event type.}

\item{learner.failure_time}{\code{binomial} \code{\link[sl3]{Lrnr_base}} learner object used for fitting the conditional hazard model for failure.}

\item{learner.censoring_time}{A \code{binomial} \code{\link[sl3]{Lrnr_base}} learner object used for fitting conditional hazard model for censoring.}

\item{placebo_risk}{Placebo risk (to calculate VE columns). If NULL (default), will not
include VE results.}

\item{placebo_se}{Placebo risk standard error (to calculate VE standard errors)
If NULL (default), will not include VE results.}
}
\value{
A data.frame containing the following columns for each threshold in `threshold_list`:
- `estimate`: The estimated value.
- `se`: The standard error of the estimate.
- `ci_lo`: The lower bound of the confidence interval.
- `ci_hi`: The upper bound of the confidence interval.
- `estimate_monotone`: The monotone-adjusted estimate.
- `ci_lo_monotone`: The lower bound of the monotone-adjusted confidence interval.
- `ci_hi_monotone`: The upper bound of the monotone-adjusted confidence interval.
- `n_in_bin`: The number of observations above the threshold.
- `n_events_in_bin`: The number of events above the threshold.
}
\description{
This method estimates \eqn{E[E[Y|A \ge v, W]]} for a range of thresholds \eqn{v}, where \eqn{Y}
is a binary outcome of interest that is subject to right-censoring, \eqn{A} is a
continuous biomarker of interest, and \eqn{W} are baseline variables. To
account for missingness in the marker \eqn{A}, the user can
provide inverse probability observation weights in the function call.
}
