% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholdSurv.R
\name{thresholdSurv}
\alias{thresholdSurv}
\title{Non-parametric estimation of the covariate-adjusted threshold-response function}
\usage{
thresholdSurv(
  data,
  covariates,
  failure_time,
  event_type,
  marker,
  weights,
  tf,
  threshold_list = NULL,
  nbins_time = 20,
  nbins_threshold = 20,
  verbose = FALSE,
  learner.treatment = NULL,
  learner.event_type = NULL,
  learner.failure_time = NULL,
  learner.censoring_time = NULL
)
}
\arguments{
\item{data}{Dataframe}

\item{covariates}{Vector of strings, representing column names for the adjustment covariates}

\item{failure_time}{A string for the column name of the failure time variable}

\item{event_type}{A string for the column name fo the event type variable}

\item{marker}{A string for the column name of the marker variable}

\item{weights}{A string for the column name of the weights used in estimation}

\item{tf}{A numeric for reference time point}

\item{threshold_list}{A vector of numeric variables of thresholds to estimate.
If NULL, then estimate on the entire dataset}

\item{nbins_time}{A numeric for time bins}

\item{nbins_threshold}{Threshold of time bins ?? (NOT USED HERE)
This learner is used as \code{binomial_learner} for \code{\link[sl3]{Lrnr_pooled_hazards}}.
Default uses auto ensemble learning by calling `causalutils::get_autoML()`.}

\item{verbose}{A logical value indicating whether to display progress and diagnostic messages during the computation. Default is FALSE.}

\item{learner.treatment}{A \code{binomial} \code{\link[sl3]{Lrnr_base}} learner object used for fitting the propensity score model for the treatment mechanism.
For nonbinary categorical treatment, this is used as \code{binomial_learner} in \code{\link[sl3]{Lrnr_independent_binomial}}.
Default uses auto ensemble learning by calling `causalutils::get_autoML()`.}

\item{learner.event_type}{A \code{binomial} \code{\link[sl3][Lrnr_base]} learner object used for fitting the conditional probability distribution for failure time of the failure event type.
For nonbinary categorical event type, this is used as \code{binomial_learner} in \code{\link[sl3]{Lrnr_independent_binomial}}.
Default uses auto ensemble learning by calling `causalutils::get_autoML()`.}

\item{learner.censoring_time}{A \code{binomial} \code{\link[sl3]{Lrnr_base}} learner object used for fitting conditional hazard model for censoring.
This learner is used as \code{binomial_learner} for \code{\link[sl3]{Lrnr_pooled_hazards}}.
Default uses auto ensemble learning by calling `causalutils::get_autoML()`.}
}
\value{
A dataframe with the estimates, estimates assuming monotonicity,
standard errors, confidence intervals, number above the threshold, number of events
above the threshold for each threshold in threshold_list.
}
\description{
https://arxiv.org/pdf/2107.11459
}
